refactor(desktop): extract markdown code block renderer
Move the markdown code block and Shiki highlighting renderer into the markdown submodule without changing its implementation.
